refactor: 보드 만들기 modal 리팩토링
